+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   - `hel*`: match the tags that contain `hel`.
 
 - Add `gtags` support for dumb_jump provider. #792
+- Introduce debounce for user typed event. #793
 
 ## Fixed